Enhance section management: refactor Section and SubSection controllers, improve course updates, and implement SubSectionModal for better handling of subsections.

+++ Ed-Tech-Backend/controllers/Section.controller.js
@@ -1,5 +1,9 @@
 const Section = require("../models/Section.models");
 const Course = require("../models/Course.models");
+const SubSection = require("../models/SubSection.models");
+
+
+
 
 exports.createSection = async (req, res) => {
     try {
@@ -15,15 +19,15 @@ exports.createSection = async (req, res) => {
         //create section
         const newSection = await Section.create({sectionName});
         //update course with section ObjectID
-        // const updatedCourseDetails = await Course.findByIdAndUpdate(
-        //                                 courseId,
-        //                                 {
-        //                                     $push:{
-        //                                         courseContent:newSection._id,
-        //                                     }
-        //                                 },
-        //                             {new:true},                                    
-        //                        )
+        await Course.findByIdAndUpdate(
+                                        courseId,
+                                        {
+                                            $push:{
+                                                courseContent:newSection._id,
+                                            }
+                                        },
+                                    {new:true},                                    
+                               )
          const returnValue = await Course.findById(courseId).populate("instructor").populate("category").populate("courseContent").exec();
         //use populate to replace sections and sub sections both in the updatedcoursedetails
         //return response
@@ -43,20 +47,25 @@ exports.createSection = async (req, res) => {
 
 exports.updateSection = async (req, res) => {
     try {
-        const {sectionName, sectionId} = req.body;
+        const {sectionName, sectionId , courseId} = req.body;
 
-        if(!sectionName || !sectionId) {
-            return res.status(400).json({
-                success:false,
-                message:"Missing Properties",
-            });
-        }
-        const section = await Section.findByIdAndUpdate(sectionId, {sectionName}, {new:true});
+        const section = await Section.findByIdAndUpdate(
+            sectionId, 
+            {sectionName}, 
+            {new:true});
+        
+        const course = await Course.findById(courseId)
+        .populate({
+            path:"courseContent",
+            populate:{
+                path:"subSection",
+            },
+        }).exec();
 
         return res.status(200).json({
             success:true,
-            message:"Section updated successfully",
-            section,
+            message: section,
+            data: course,
         })
     } catch (error) {
         console.error(error);
@@ -69,12 +78,34 @@ exports.updateSection = async (req, res) => {
 
 exports.deleteSection = async (req, res) => {
     try {
-        const {sectionId} = req.body;
+        const {sectionId , courseId} = req.body;
+        await Course.findByIdAndUpdate(courseId , {
+             $pull: {
+                courseContent: sectionId,
+             }
+        });
+        const section = await Section.findById(sectionId);
+        if(!section) {
+            return res.status(404).json({
+                success:false,
+                message:"Section not found",
+            });
+        }
+        //delete subsection
+        await SubSection.deleteMany({_id:{$in:section.subSection}});
         await Section.findByIdAndDelete(sectionId);
+        //find the updated course and return
+        const course = await Course.findById(courseId).populate({
+            path:"courseContent",
+            populate: {
+                path: "subSection"
+            }
+        }).exec();            
         //Testing: do we need to delete the entry from the course schema?
         return res.status(200).json({
             success:true,
             message:"Section deleted successfully",
+            data: course,
         })
     } catch (error) {
         return res.status(500).json({
